Fixes for vl3_dns
